# back_end/v2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "没有描述",
        "title": "超市进销存系统--数据库课设接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/producer/delete": {
            "post": {
                "description": "删除供应商，返回外键检查结果和删除结果，如果有外键依赖则会提示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "producer"
                ],
                "summary": "删除供应商",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Producer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/producer/insert": {
            "post": {
                "description": "批量导入供应商，并且返回成功失败列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "producer"
                ],
                "summary": "批量导入供应商",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProducerList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProducerInsertResponseDto"
                        }
                    }
                }
            }
        },
...
        "/producer/update": {
            "post": {
                "description": "更新供应商，除了id之外都能改，所有不会有外键冲突",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "producer"
                ],
                "summary": "更新供应商",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Producer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/product/delete": {
            "post": {
                "description": "删除产品，返回外键检查结果和删除结果，如果有外键依赖则会提示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "删除产品",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/insert": {
            "post": {
                "description": "批量导入产品，并且返回成功失败列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "批量导入产品",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ProductList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductInsertResponse"
                        }
                    }
                }
            }
        },
...
        "/product/update": {
            "post": {
                "description": "更新产品，除了id之外都能改，所有不会有外键冲突",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "更新产品",
                "parameters": [
                    {
                        "description": "Producer Request",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ProducerInsertResponseDto": {
            "type": "object",
            "properties": {
                "fail_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Producer"
                    }
                },
                "success_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Producer"
                    }
                }
            }
        },
        "dto.ProducerList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Producer"
                    }
                }
            }
        },
        "dto.ProductInsertResponse": {
            "type": "object",
            "properties": {
                "fail_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                },
                "success_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                }
            }
        },
        "dto.ProductList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Product"
                    }
                }
            }
        },
...
        "models.Producer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "contact_name": {
                    "type": "string"
                },
                "contact_tel": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "short_name": {
                    "type": "string"
                },
                "tel": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "introduction": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "producer_id": {
                    "type": "integer"
                }
            }
        }
    }
}